requirements.txt
streamlit
moviepy
edge-tts
pillow
numpy
